Add configurable time source and connection cleanup
- Changed LastSeen from uint32 to int64 for better time representation
- Added Now func() int64 field to Table for customizable time source
- Default to time.Now().Unix() but allow override for performance
- Added Cleanup(now int64) method to NAT interface
- Implement automatic cleanup with protocol-specific timeouts:
  - TCP: 24 hours
  - UDP: 3 minutes
  - ICMP: 30 seconds
- Updated README with cleanup example and performance optimization
- Added example showing how to use custom time source

This allows users to optimize time.Now() calls in high-performance
scenarios while maintaining proper connection tracking cleanup.

+++ pair.go
@@ -65,7 +65,7 @@ func (p *Pair[IP]) removeConnection(conn *Conn[IP]) {
 	delete(p.in, externalKey)
 }
 
-func (p *Pair[IP]) cleanupExpired(now uint32, timeout uint32) {
+func (p *Pair[IP]) cleanupExpired(now int64, timeout int64) {
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
